+++ Frontend/src/assets/Footer/sies_logo_footer.png
